pyproject.toml
[project]
name = "sgp-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "atproto>=0.0.61",
    "python-dotenv>=1.1.1",
]
